+++ Cargo.toml
@@ -2,6 +2,7 @@
 name = "concread"
 version = "0.1.20"
 authors = ["William Brown <william@blackhats.net.au>"]
+edition = "2018"
 
 description = "Concurrently Readable Data-Structures for Rust"
 documentation = "https://docs.rs/concread/latest/concread/"
@@ -14,6 +15,7 @@ license = "MPL-2.0"
 
 [features]
 simd_support = ["packed_simd"]
+asynch = ["tokio"]
 skinny = []
 
 [dependencies]
@@ -26,9 +28,11 @@ smallvec = "1.4"
 ahash = "0.3"
 rand = "0.7"
 packed_simd = { version = "0.3", optional = true }
+tokio = { version = "0.2", optional = true, features = ["sync"] }
 
 [dev-dependencies]
 time = "0.2"
+async-std = "1.6"
 
 [[bin]]
 name = "unsound"
